+++ packages/icons/ArrowheadRight.svelte
@@ -1,9 +1,0 @@
-<!--
-This icon was adapted from the Photon Design System (licensed under the MPL v2.0):
-https://design.firefox.com/photon/
-https://github.com/FirefoxUX/photon-icons/blob/master/icons/desktop/arrowhead-right-16.svg
--->
-<script>
-export let size = 24;
-</script>
-<svg xmlns="http://www.w3.org/2000/svg" width={size} height={size} viewBox="0 0 16 16"><path fill="currentColor" d="M6 14a1 1 0 0 1-.707-1.707L9.586 8 5.293 3.707a1 1 0 0 1 1.414-1.414l5 5a1 1 0 0 1 0 1.414l-5 5A1 1 0 0 1 6 14z"></path></svg>