+++ packages/icons/Forward.svelte
@@ -1,9 +1,0 @@
-<!--
-This icon was adapted from the Photon Design System (licensed under the MPL v2.0):
-https://design.firefox.com/photon/
-https://github.com/FirefoxUX/photon-icons/blob/master/icons/desktop/forward-16.svg
--->
-<script>
-export let size = 24;
-</script>
-<svg xmlns="http://www.w3.org/2000/svg" width={size} height={size} viewBox="0 0 16 16"><path fill="currentColor" d="M15.707 7.293l-6-6a1 1 0 0 0-1.414 1.414L12.586 7H1a1 1 0 0 0 0 2h11.586l-4.293 4.293a1 1 0 1 0 1.414 1.414l6-6a1 1 0 0 0 0-1.414z"></path></svg>