+++ packages/icons/Table.svelte
@@ -1,10 +1,11 @@
 <script>
 export let size = 24;
+export let color = 'currentColor';
 </script>
 
 
 <svg width={size} height={size} viewBox="0 0 24 24" version="1.1" xmlns="http://www.w3.org/2000/svg" xmlns:xlink="http://www.w3.org/1999/xlink" xml:space="preserve" xmlns:serif="http://www.serif.com/" style="fill-rule:evenodd;clip-rule:evenodd;stroke-linejoin:round;stroke-miterlimit:2;">
-    <g fill=currentColor>
+    <g fill={color}>
       <path d="M20.043,2C21.124,2 22,2.876 22,3.957L22,20C22,21.104 21.104,22 20,22L3.957,22C3.438,22 2.94,21.794 2.573,21.427C2.206,21.06 2,20.562 2,20.043L2,3.946C2,2.871 2.871,2 3.946,2C7.767,2 16.213,2 20.043,2ZM20,6L4,6L4,20L20,20L20,6Z"/>
       <path d="M4,12L20,12L20,14L4,14L4,12"/>
       <rect x="11" y="6" width="2" height="14"/>