+++ packages/icons/Undo.svelte
@@ -2,8 +2,6 @@
 export let size = 24;
 export let color = 'currentColor';
 </script>
-
 <svg width={size} height={size} fill={color} viewBox="0 0 18 18" version="1.1" xmlns="http://www.w3.org/2000/svg" xmlns:xlink="http://www.w3.org/1999/xlink" xml:space="preserve" xmlns:serif="http://www.serif.com/" style="fill-rule:evenodd;clip-rule:evenodd;stroke-linejoin:round;stroke-miterlimit:2;">
-  <path d="M17,3.013C17,1.901 16.099,1 14.987,1L3.03,1C2.492,1 1.975,1.214 1.595,1.595C1.214,1.975 1,2.492 1,3.03L1,15.01C1,15.538 1.21,16.044 1.583,16.417C1.956,16.79 2.462,17 2.99,17C6.062,17 11.925,17 15.002,17C15.532,17 16.04,16.789 16.415,16.415C16.789,16.04 17,15.532 17,15.002C17,11.932 17,6.091 17,3.013ZM15,3L3,3L3,15L15,15L15,3Z"/>
+  <path d="M4.933,8.406L2.5,6L2.5,12L8.566,12L6.16,9.621C6.955,8.715 8.126,8.143 9.432,8.143C11.677,8.143 13.525,9.835 13.743,12L15.482,12C15.26,8.888 12.634,6.429 9.432,6.429C7.648,6.429 6.043,7.192 4.933,8.406Z"/>
 </svg>
-