+++ packages/icons/base.js
@@ -1,43 +1,113 @@
-import ArrowheadDown from "./ArrowheadDown.svelte";
-import ArrowheadLeft from "./ArrowheadLeft.svelte";
-import ArrowheadRight from "./ArrowheadRight.svelte";
-import ArrowheadUp from "./ArrowheadUp.svelte";
 import BarChart from "./BarChart.svelte";
+import Calendar from "./Calendar.svelte";
 import Cancel from "./Cancel.svelte";
 import Check from "./Check.svelte";
 import Checkbox from "./Checkbox.svelte";
 import CheckboxBlank from "./CheckboxBlank.svelte";
 import CheckboxBlankDark from "./CheckboxBlankDark.svelte";
 import Close from "./Close.svelte";
-import Forward from "./Forward.svelte";
 import Help from "./Help.svelte";
 import Info from "./Info.svelte";
 import Information from "./Information.svelte";
 import LineChart from "./LineChart.svelte";
+import Density from "./Density.svelte";
 import Menu from "./Menu.svelte";
 import Search from "./Search.svelte";
 import Table from "./Table.svelte";
 import Warning from "./Warning.svelte";
+import Add from "./Add.svelte";
+import AddCircle from "./AddCircle.svelte";
+import AddCircleDark from "./AddCircleDark.svelte";
+import AddDark from "./AddDark.svelte";
+
+import Tile from "./Tile.svelte";
+import Stack from "./Stack.svelte";
+
+import Undo from "./Undo.svelte";
+import Redo from "./Redo.svelte";
+
+import RadioButton from "./RadioButton.svelte";
+import RadioButtonEmpty from "./RadioButtonEmpty.svelte";
+
+import Remove from "./Remove.svelte";
+import RemoveCircle from "./RemoveCircle.svelte";
+import RemoveCircleDark from "./RemoveCircleDark.svelte";
+import RemoveDark from "./RemoveDark.svelte";
+import MoreHorizontal from "./MoreHorizontal.svelte";
+import MoreVertical from "./MoreVertical.svelte";
+import Sequence from "./Sequence.svelte";
+
+import Documentation from "./Documentation.svelte";
+import DocumentationDark from "./DocumentationDark.svelte";
+
+import Code from "./Code.svelte";
+
+import CaretUp from "./CaretUp.svelte";
+import CaretDown from "./CaretDown.svelte";
+import CaretLeft from "./CaretLeft.svelte";
+import CaretRight from "./CaretRight.svelte";
+import FirstPage from "./FirstPage.svelte";
+import LastPage from "./LastPage.svelte";
+
+import ArrowUp from "./ArrowUp.svelte";
+import ArrowDown from "./ArrowDown.svelte";
+import ArrowLeft from "./ArrowLeft.svelte";
+import ArrowRight from "./ArrowRight.svelte";
+import ArrowUpLeft from "./ArrowUpLeft.svelte";
+import ArrowUpRight from "./ArrowUpRight.svelte";
+import ArrowDownLeft from "./ArrowDownLeft.svelte";
+import ArrowDownRight from "./ArrowDownRight.svelte";
 
 export {
-  ArrowheadDown,
-  ArrowheadLeft,
-  ArrowheadRight,
-  ArrowheadUp,
+  Sequence,
+  LineChart,
   BarChart,
+  Density,
+  Table,
+  Calendar,
+  Code,
+  Tile,
+  Stack,
+  Add,
+  AddDark,
+  AddCircle,
+  AddCircleDark,
+  Remove,
+  RemoveDark,
+  RemoveCircle,
+  RemoveCircleDark,
+  RadioButton,
+  RadioButtonEmpty,
+  CaretUp,
+  CaretDown,
+  CaretLeft,
+  CaretRight,
+  FirstPage,
+  LastPage,
+  ArrowUp,
+  ArrowDown,
+  ArrowLeft,
+  ArrowRight,
+  ArrowUpLeft,
+  ArrowUpRight,
+  ArrowDownLeft,
+  ArrowDownRight,
+  Undo,
+  Redo,
+  Documentation,
+  DocumentationDark,
   Cancel,
   Check,
   Checkbox,
   CheckboxBlank,
   CheckboxBlankDark,
   Close,
-  Forward,
   Help,
   Info,
   Information,
-  LineChart,
+  MoreHorizontal,
+  MoreVertical,
   Menu,
   Search,
-  Table,
   Warning,
 };